+++ packages/svelte-undp-components/src/lib/components/maplibre/line/LineColor.stories.svelte
@@ -299,5 +299,5 @@
 		colorMapName: 'viridis',
 		defaultColor: '#000000'
 	}}
-	children={template}
+	{template}
 />