+++ packages/svelte-undp-components/src/lib/components/maplibre/line/LineWidth.stories.svelte
@@ -302,5 +302,5 @@
 		classificationMethod: 'n',
 		numberOfRandomSamplingPoints: 1000
 	}}
-	children={template}
+	{template}
 />