+++ packages/svelte-undp-components/src/lib/components/maplibre/symbol/TextSize.stories.svelte
@@ -46,4 +46,4 @@
 	</Map>
 {/snippet}
 
-<Story name="Primary" args={{ layerId: layer.id }} children={template} />
+<Story name="Primary" args={{ layerId: layer.id }} {template} />