+++ packages/svelte-undp-components/src/lib/components/titiler/RasterClassifyLegend.stories.svelte
@@ -126,5 +126,5 @@
 		colorMapName: 'brbg',
 		numberOfClasses: 6
 	}}
-	children={template}
+	{template}
 />