+++ sites/geohub/src/components/pages/map/data/RasterAlgorithmExplorer.svelte
@@ -63,7 +63,7 @@
 	}
 
 	let {
-		feature = $bindable(),
+		feature,
 		title = 'Choose a tool',
 		cardDescription = 'use this tool',
 		mode = 'map',