+++ sites/geohub/src/lib/server/api/datasets/[id]/style/[layer]/[type]/DELETE.ts
@@ -14,8 +14,8 @@ export const Param = z.object({
 	id: z.string().describe('Dataset ID'),
 	layer: z.string().describe('Band name if it is raster, layer ID if it is vector.'),
 	type: z
-		.enum(['raster', 'fill', 'symbol', 'line', 'circle', 'heatmap'])
-		.describe('Maplibre layer type (fill, line, symbol, circle, heatmap, raster)')
+		.enum(['raster', 'fill', 'symbol', 'line', 'circle', 'heatmap', 'fill-extrusion'])
+		.describe('Maplibre layer type (fill, line, symbol, circle, heatmap, raster, fill-extrusion)')
 });
 
 const description = `