+++ packages/svelte-undp-components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @undp-data/svelte-undp-components
 
+## 2.0.10
+
+### Patch Changes
+
+- 2d81062: chore: updated dependencies
+- Updated dependencies [2d81062]
+  - @undp-data/svelte-undp-design@3.0.6
+
 ## 2.0.9
 
 ### Patch Changes